feat: moved Ridge from preview to main sklearnex
